The assets are in, placed in and working, pathfinding is borked...
If I'm being entirely honest, I kinda didn't focus on this game too much had life in the way and other assessments (This isn't an excuse) I'm sorry but this is as much as I could do and I'll submit it regardless of its finish product. I thank everyone who still put effort towards the game especially the art team fantastic job! Programming was left to be desired... Thanks Jayden for doing most of the functionality I ended up doing the UI

Co-Authored-By: JaylanMuyesser <60990478+JaylanMuyesser@users.noreply.github.com>
Co-Authored-By: kypilot <69182478+kypilot@users.noreply.github.com>
Co-Authored-By: georgia-pinkerton <79297318+georgia-pinkerton@users.noreply.github.com>